Хочу пелеменей

+++ VkLogic/src/main/java/ru/spliterash/vkchat/db/dao/ConversationDao.java
@@ -6,7 +6,7 @@ import ru.spliterash.vkchat.db.model.ConversationModel;
 
 import java.sql.SQLException;
 
-public class ConversationDao extends BaseDaoImpl<ConversationModel, Long> {
+public class ConversationDao extends BaseDaoImpl<ConversationModel, Integer> {
     protected ConversationDao(ConnectionSource source) throws SQLException {
         super(source,ConversationModel.class);
     }
